fix(app): follow the BigNumber wallet through the app and storage
The previous commit moved `Resources.ore` to `BigNumber` in Core and left the
app unbuildable. This carries it the rest of the way.

Storage keeps the legacy `ore: Double` column and adds `oreMantissa` /
`oreExponent`. A row where both are zero predates the migration and is read from
the old column, so existing saves open unchanged and new writes fill all three.

`DeepMineNumberFormatter` gains a `string(big:)` entry point rather than an
overload. That is deliberate: `BigNumber` is `ExpressibleBy*Literal`, so an
overload makes every literal call site ambiguous — the third time that trap
appeared in this migration, after `Resources(ore:)` and `BigNumber`'s own
comparison operators. Values past 1e15 print in scientific notation, because
Korean has no everyday word above 조 and an unbounded wallet passes it routinely.

Equipment rows drop the "maxed out" state: with no ceiling, depth is the only
thing that can gate a purchase.

Verification:
  xcodebuild test -only-testing:DeepMineAppTests → 134/134
  xcodebuild -destination 'generic/platform=iOS' CODE_SIGNING_ALLOWED=NO → clean

+++ DeepMineApp/Session/GameStore+SessionPresentation.swift
@@ -245,7 +245,7 @@ extension GameStore {
                     currentLevel: level,
                     nextLevel: level + 1,
                     cost: cost,
-                    availableOre: player.resources.ore,
+                    availableOre: player.resources.ore.doubleValue,
                     marginalExpectedOre: gain
                 )
             }
